requirements.txt
google-adk==1.14.0
langchain-community==0.3.27
wikipedia==1.4.0
google-generativeai==0.8.5